# cc2-openapi.yaml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Microservices CC2
  license:
    name: MIT
servers:
  - url: http://microservices.cc2.io/v1
paths:
  /payments:
    post:
      summary: Initiate a payment process
      operationId: initiatePaymentProcess
      tags:
        - payments 
      requestBody:
        description: Body of a request to start a payment process 
        required: true 
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentProcess"
      responses:
        '202':
          description: Request accepted
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    BuyerInfo:
      type: object
      required:
        - id
        - fullname
      properties:
        id:
          type: string
        fullname:
          type: string
      
    CreditCardInfo:
      type: object 
      required:
        - provider
        - token
      properties:
        provider:
          type: string 
        token:
          type: string

    PaymentOrder:
      type: object
      required: 
        - seller_account
        - amount 
        - currency
        - payment_order_id
      properties:
        seller_account:
          type: string
        amount:
          type: string 
        currency:
          type: string 
        payment_order_id:
          type: string

    InitiatePaymentProcess:
      type: object
      required:
        - buyer_info
        - checkout_id
        - credit_card_info
        - payment_orders
      properties:
        buyer_info:
          $ref: "#/components/schemas/BuyerInfo"
        checkout_id: 
          type: string 
        credit_card_info:
          $ref: "#/components/schemas/CreditCardInfo"
        payment_orders:
          type: array
          items:
            $ref: "#/components/schemas/PaymentOrder"

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string